requirements.txt
flask>=2.0
flask-sqlalchemy>=3.0
flask-login>=0.6
python-twitter>=3.5
gunicorn>=21.0
